refactor(app): move some ot2 cal types around
